+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 When making changes, add an entry to the **Unreleased** section.
 
-## [Unreleased]
+## [0.8.3] - 2025-07-11
 
 ### Changed